steps is now int

+++ draw.h
@@ -30,13 +30,13 @@ struct matrix * generate_torus( double cx, double cy, double cz,
 //2D Curves
 void add_circle(struct matrix * points,
                 double cx, double cy, double cz,
-                double r, double step );
+                double r, int step );
 void add_curve( struct matrix *points,
                 double x0, double y0,
                 double x1, double y1,
                 double x2, double y2,
                 double x3, double y3,
-                double step, int type );
+                int step, int type );
 
 void add_point(struct matrix * points, double x, double y, double z);
 void add_edge(struct matrix * points,
